refactor: remove ActionCreatedAndProcessed fixture method and update tests
