+++ pkg/storage/internal/testabilities/fixture_storage.go
@@ -1,25 +1,18 @@
 package testabilities
 
 import (
+	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/internal/fixtures"
 	"testing"
 
-	"github.com/bsv-blockchain/go-sdk/transaction"
-	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/internal/fixtures"
 	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/internal/fixtures/testusers"
 	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/internal/testabilities"
 	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/storage"
-	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/storage/internal/testabilities/tsgenerated"
 	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/wdk"
-	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/wdk/primitives"
-	"github.com/go-softwarelab/common/pkg/to"
-	"github.com/stretchr/testify/require"
 )
 
 type StorageFixture interface {
 	testabilities.StorageFixture
 	StorageManagerForUser(user testusers.User, activeStorage wdk.WalletStorageProvider) *storage.WalletStorageManager
-	ActionCreatedAndSigned(activeStorage *storage.Provider) (createActionResult *wdk.StorageCreateActionResult, signedTransaction *transaction.Transaction)
-	ActionCreatedAndProcessed(activeStorage *storage.Provider) (createActionResult *wdk.StorageCreateActionResult, signedTransaction *transaction.Transaction)
 	Action(activeStorage *storage.Provider) TxGeneratorFixture
 }
 
@@ -32,98 +25,16 @@ func (s *storageFixture) StorageManagerForUser(user testusers.User, activeStorag
 	return storage.NewWalletStorageManager(user.IdentityKey(s.t), activeStorage)
 }
 
-func (s *storageFixture) ActionCreatedAndSigned(activeStorage *storage.Provider) (createActionResult *wdk.StorageCreateActionResult, signedTransaction *transaction.Transaction) {
-	ctx := s.t.Context()
-	internalizeArgs := wdk.InternalizeActionArgs{
-		Tx: tsgenerated.AtomicBeefToInternalize(s.t),
-		Outputs: []*wdk.InternalizeOutput{
-			{
-				OutputIndex: 0,
-				Protocol:    wdk.WalletPaymentProtocol,
-				PaymentRemittance: &wdk.WalletPayment{
-					DerivationPrefix:  fixtures.DerivationPrefix,
-					DerivationSuffix:  fixtures.DerivationSuffix,
-					SenderIdentityKey: fixtures.AnyoneIdentityKey,
-				},
-			},
-		},
-		Description: "description",
+func (s *storageFixture) Action(activeStorage *storage.Provider) TxGeneratorFixture {
+	return &txGeneratorFixture{
+		TB:                    s.t,
+		satoshisToInternalize: fixtures.DefaultCreateActionOutputSatoshis,
+		satoshisToSend:        1,
+		parent:                s,
+		activeStorage:         activeStorage,
+		sender:                testusers.Alice,
+		recipient:             testusers.Bob,
 	}
-
-	// NOTE: Alice's identityKey has been used for tsgenerated.SignedTransaction - that's why you cannot use another user here
-	user := testusers.Alice
-
-	s.Provider().BHS().OnMerkleRootVerifyResponse(
-		tsgenerated.BeefToInternalizeHeight,
-		tsgenerated.BeefToInternalizeMerkleRoot,
-		BHSMerkleRootConfirmed,
-	)
-
-	_, err := activeStorage.InternalizeAction(ctx, user.AuthID(), internalizeArgs)
-	require.NoError(s.t, err)
-
-	args := wdk.ValidCreateActionArgs{
-		Description: "outputBRC29",
-		Inputs:      []wdk.ValidCreateActionInput{},
-		Outputs: []wdk.ValidCreateActionOutput{
-			{
-				LockingScript:      "76a9144b0d6cbef5a813d2d12dcec1de2584b250dc96a388ac",
-				Satoshis:           1000,
-				OutputDescription:  "outputBRC29",
-				CustomInstructions: to.Ptr(`{"derivationPrefix":"Pr==","derivationSuffix":"Su==","type":"BRC29"}`),
-				Tags:               []primitives.StringUnder300{fixtures.CreateActionTestTag},
-			},
-		},
-		LockTime: 0,
-		Version:  1,
-		Labels:   []primitives.StringUnder300{fixtures.CreateActionTestLabel},
-		Options: wdk.ValidCreateActionOptions{
-			AcceptDelayedBroadcast: to.Ptr[primitives.BooleanDefaultTrue](false),
-			SendWith:               []primitives.TXIDHexString{},
-			SignAndProcess:         to.Ptr(primitives.BooleanDefaultTrue(true)),
-			KnownTxids:             []primitives.TXIDHexString{},
-			NoSendChange:           []wdk.OutPoint{},
-			RandomizeOutputs:       false,
-		},
-		IsSendWith:                   false,
-		IsDelayed:                    false,
-		IsNoSend:                     false,
-		IsNewTx:                      true,
-		IsRemixChange:                false,
-		IsSignAction:                 false,
-		IncludeAllSourceTransactions: true,
-	}
-
-	result, err := activeStorage.CreateAction(
-		ctx,
-		testusers.Alice.AuthID(),
-		args,
-	)
-
-	require.NoError(s.t, err)
-
-	return result, tsgenerated.SignedTransaction(s.t)
-}
-
-func (s *storageFixture) ActionCreatedAndProcessed(activeStorage *storage.Provider) (createActionResult *wdk.StorageCreateActionResult, signedTransaction *transaction.Transaction) {
-	s.t.Helper()
-	createActionResult, signedTx := s.ActionCreatedAndSigned(activeStorage)
-	txID := signedTx.TxID().String()
-
-	args := wdk.ProcessActionArgs{
-		IsNewTx:    true,
-		IsSendWith: false,
-		IsNoSend:   false,
-		IsDelayed:  false,
-		Reference:  to.Ptr(createActionResult.Reference),
-		TxID:       to.Ptr(primitives.TXIDHexString(txID)),
-		RawTx:      signedTx.Bytes(),
-		SendWith:   []primitives.TXIDHexString{},
-	}
-
-	_, err := activeStorage.ProcessAction(s.t.Context(), testusers.Alice.AuthID(), args)
-	require.NoError(s.t, err)
-	return createActionResult, signedTx
 }
 
 func Given(t testing.TB) (given StorageFixture, cleanup func()) {
